+++ TestServer/Common/Network/Packets/MessageInfo.cs
