+++ TestServer/Common/Network/Protocols/TcpClient.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Common.Network.Protocols
+﻿namespace Common.Network.Protocols
 {
     class TcpClient
     {
