+++ TestServer/Server/Db/RemoveClientFromChat.cs
@@ -1,14 +1,14 @@
-﻿namespace Common.Network
+﻿namespace Server.DataBase
 {
     using System.Collections.Generic;
-    public class CreatingChatInfo
+    public class RemoveClientFromChat
     {
         #region Properties
 
-        public string NameOfClientSender { get; set; }
-
+        public int NumberChat { get; set; }
         public List<string> NameOfClients { get; set; }
 
         #endregion Properties
+
     }
 }
