+++ TestServer/Server/DbEF/PoolClients.cs
@@ -1,4 +1,4 @@
-﻿namespace Common.DbEF
+﻿namespace Server.DataBase
 {
     using System.ComponentModel.DataAnnotations;
     using System.Collections.Generic;
