+++ TestServer/Server/Factory/FactoryTransport.cs
@@ -1,5 +1,7 @@
-﻿namespace Common.Network
+﻿namespace Server.Network
 {
+    using Common.Network;
+    using Configuration;
     using System.Net;
     using System;
     public static class TransportFactory
@@ -10,9 +12,9 @@
         {
             switch (config.Protocol)
             {
-                case TransportType.WebSocket:
+                case TypeTransport.WebSocket:
                     return new WsServer(new IPEndPoint(IPAddress.Any, config.Port));
-                case TransportType.Tcp:
+                case TypeTransport.Tcp:
                     return new TcpServer(new IPEndPoint(IPAddress.Any, config.Port));
                 default:
                     throw new ArgumentOutOfRangeException(nameof(config), config, null);
