+++ TestServer/Server/NetworkManager.cs
@@ -1,6 +1,10 @@
 ﻿namespace Server
 {
+	using Server.Network;
+	using Common.Configuration;
 	using Common.Network;
+	using Configuration;
+	using DataBase;
 	class NetworkManager
 	{
